+++ openghg_inversions/postprocessing/diagnostics.py
@@ -6,7 +6,11 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from openghg_inversions.postprocessing.inversion_output import PostprocessingInput, as_postprocessing_output
+from openghg_inversions.postprocessing.inversion_output import (
+    InversionOutput,
+    standard_obs_inputs,
+    standard_trace_dataset,
+)
 from openghg_inversions.postprocessing.utils import add_suffix, get_parameters
 
 Diagnostic = namedtuple("Diagnostic", ["func", "params"])
@@ -30,7 +34,7 @@ def register_diagnostic(diagnostic: Callable) -> Callable:
 
 @register_diagnostic
 @add_suffix("trace")
-def summary(inv_out: PostprocessingInput) -> xr.Dataset:
+def summary(inv_out: InversionOutput) -> xr.Dataset:
     """Return diagnostics summary computed by arviz.
 
     Diagnostics reported:
@@ -104,7 +108,7 @@ def _r2_by_site(ds: xr.Dataset, report_prior: bool = False) -> xr.Dataset:
 
 
 @register_diagnostic
-def bayes_r2_by_site(inv_out: PostprocessingInput, report_prior: bool = False) -> xr.Dataset:
+def bayes_r2_by_site(inv_out: InversionOutput, report_prior: bool = False) -> xr.Dataset:
     """Compute Bayesian R2 scores grouped by site.
 
     Scores are computed for posterior predictive traces (compared
@@ -123,9 +127,8 @@ def bayes_r2_by_site(inv_out: PostprocessingInput, report_prior: bool = False) -
             with uncertainties.
 
     """
-    inv_out = as_postprocessing_output(inv_out)
-    y_true = inv_out.obs_inputs["y_obs"].unstack("nmeasure")
-    y_pred = inv_out.get_trace_dataset(var_names="y").unstack("nmeasure")
+    y_true = standard_obs_inputs(inv_out)["y_obs"].unstack("nmeasure")
+    y_pred = standard_trace_dataset(inv_out, var_names="y").unstack("nmeasure")
     ds = xr.merge([y_true, y_pred])
 
     return _r2_by_site(ds, report_prior=report_prior)
@@ -133,7 +136,7 @@ def bayes_r2_by_site(inv_out: PostprocessingInput, report_prior: bool = False) -
 
 @register_diagnostic
 def bayes_r2_by_site_resample(
-    inv_out: PostprocessingInput, freq: str = "MS", report_prior: bool = False
+    inv_out: InversionOutput, freq: str = "MS", report_prior: bool = False
 ) -> xr.Dataset:
     """Compute Bayesian R2 scores grouped by site and time.
 
@@ -155,9 +158,8 @@ def bayes_r2_by_site_resample(
             with uncertainties.
 
     """
-    inv_out = as_postprocessing_output(inv_out)
-    y_true = inv_out.obs_inputs["y_obs"].unstack("nmeasure")
-    y_pred = inv_out.get_trace_dataset(var_names="y").unstack("nmeasure")
+    y_true = standard_obs_inputs(inv_out)["y_obs"].unstack("nmeasure")
+    y_pred = standard_trace_dataset(inv_out, var_names="y").unstack("nmeasure")
     ds = xr.merge([y_true, y_pred])
 
     results = []